Fix import on client, make sure sources work on sdk,  and make sure search test works properly
