+++ nucliadb/nucliadb/search/api/v1/resource.py
@@ -28,7 +28,7 @@ from grpc.aio import AioRpcError  # type: ignore
 from nucliadb_protos.nodereader_pb2 import ParagraphSearchResponse
 from sentry_sdk import capture_exception
 
-from nucliadb.ingest.serialize import ResourceProperties, get_resource_uuid_by_slug
+from nucliadb.ingest.serialize import get_resource_uuid_by_slug
 from nucliadb.search import SERVICE_NAME, logger
 from nucliadb.search.search.fetch import abort_transaction
 from nucliadb.search.search.merge import merge_paragraphs_results
@@ -40,6 +40,7 @@ from nucliadb_models.common import FieldTypeName
 from nucliadb_models.resource import ExtractedDataTypeName, NucliaDBRoles
 from nucliadb_models.search import (
     NucliaDBClientType,
+    ResourceProperties,
     ResourceSearchResults,
     SearchOptions,
     SortOption,
