+++ nucliadb/nucliadb/search/api/v1/suggest.py
@@ -29,7 +29,6 @@ from nucliadb_protos.nodereader_pb2 import SuggestResponse
 from nucliadb_protos.writer_pb2 import ShardObject
 from sentry_sdk import capture_exception
 
-from nucliadb.ingest.serialize import ResourceProperties
 from nucliadb.search import logger
 from nucliadb.search.api.v1.router import KB_PREFIX, api
 from nucliadb.search.search.fetch import abort_transaction  # type: ignore
@@ -43,6 +42,7 @@ from nucliadb_models.resource import NucliaDBRoles
 from nucliadb_models.search import (
     KnowledgeboxSuggestResults,
     NucliaDBClientType,
+    ResourceProperties,
     SuggestOptions,
 )
 from nucliadb_utils.authentication import requires
