+++ nucliadb_sdk/requirements-sources.txt
@@ -1,0 +1,1 @@
+-e ./nucliadb_models
